requirements.txt
tensorflow
streamlit
streamlit-drawable-canvas
numpy
Pillow
